+++ test/state/HypermediaState.test.js
@@ -1,6 +1,6 @@
 
-import { assert, expect } from '@open-wc/testing';
-import { clearStore, dispose, processRawJsonSirenEntity, stateFactory } from '../../state/HypermediaState.js';
+import { dispose, processRawJsonSirenEntity, stateFactory } from '../../state/HypermediaState.js';
+import { assert } from '@open-wc/testing';
 import { fetch } from '../../state/fetch.js';
 import { FetchError } from '../../state/Fetchable.js';
 import fetchMock from 'fetch-mock/esm/client.js';
@@ -158,7 +158,7 @@ describe('HypermediaState class', () => {
 			state.addObservables(observer, observable);
 
 			await fetch(state);
-			await state.allFetchesComplete;
+			await state.allFetchesComplete();
 
 			assert.isTrue(mock.called(entityHref));
 			assert.isTrue(mock.called(linkedHref));
@@ -523,153 +523,6 @@ describe('HypermediaState class', () => {
 			const newState = window.D2L.Foundation.StateStore.get(href, rawToken);
 			assert.equal(newState.entityID, href);
 			assert.equal(newState.token.rawToken, rawToken);
-		});
-	});
-
-	describe('should know when the tree has loaded', () => {
-		const parentHref = `http://pink-${uniqueId()}`;
-		const subEntityFirst = `${parentHref}/sub1'`;
-		const subEntityFirstSub = `${subEntityFirst}/sub2'`;
-		const subEntitySecond = `${parentHref}/sub2'`;
-		const primeLink = `${parentHref}/prime`;
-		const primeLinkSubEntity = `${primeLink}/sub`;
-
-		beforeEach(async() => {
-			const parentJson = JSON.stringify({
-				class: [ 'pink' ],
-				properties: {
-					answer: 42
-				},
-				entities: [{
-					rel: [ 'first' ],
-					href: subEntityFirst
-				}, {
-					rel: [ 'second' ],
-					href: subEntitySecond
-				}],
-				links: [
-					{ rel: [ 'self' ], href: parentHref },
-					{ rel: [ 'prime' ], href: primeLink },
-				]
-			});
-
-			const subEntityFirstJson = JSON.stringify({
-				properties: {
-					answer: 43
-				},
-				entities: [{
-					rel: [ 'third' ],
-					href: subEntityFirstSub
-				}],
-				links: [
-					{ rel: [ 'self' ], href: subEntityFirstSub },
-				]
-			});
-
-			const subEntityFirstSubJson = JSON.stringify({
-				properties: {
-					answer: 44
-				},
-				links: [
-					{ rel: [ 'self' ], href: subEntityFirstSub },
-				]
-			});
-
-			const subEntitySecondJson = JSON.stringify({
-				properties: {
-					answer: 45
-				},
-				links: [
-					{ rel: [ 'self' ], href: subEntitySecond },
-				]
-			});
-
-			const primeLinkJson = JSON.stringify({
-				properties: {
-					answer: 46
-				},
-				entities: [{
-					rel: [ 'forth' ],
-					href: primeLinkSubEntity
-				}],
-				links: [
-					{ rel: [ 'self' ], href: primeLink },
-				]
-			});
-
-			const primeLinkSubEntityJson = JSON.stringify({
-				properties: {
-					answer: 47
-				},
-				links: [
-					{ rel: [ 'self' ], href: primeLinkSubEntity },
-				]
-			});
-			const option = { delay: 20 };
-			fetchMock
-				.mock(parentHref, parentJson, option)
-				.mock(subEntityFirst, subEntityFirstJson, option)
-				.mock(subEntityFirstSub, subEntityFirstSubJson, option)
-				.mock(subEntitySecond, subEntitySecondJson, option)
-				.mock(primeLink, primeLinkJson, option)
-				.mock(primeLinkSubEntity, primeLinkSubEntityJson, option);
-		});
-
-		afterEach(() => {
-			fetchMock.reset();
-			clearStore();
-		});
-
-		it('the parent state, should know when the children loaded', async() => {
-			const state = await stateFactory(parentHref, 'token');
-			const observable = {
-				parentAnswer: {
-					observable: observableTypes.property, id: 'answer'
-				},
-				subEntityFirstAnswer: {
-					observable: observableTypes.property, id: 'answer',
-					route: [{ observable: observableTypes.subEntity, rel: 'first' }]
-				},
-				subEntityFirstSubAnswer: {
-					observable: observableTypes.property, id: 'answer',
-					route: [
-						{ observable: observableTypes.subEntity, rel: 'first' },
-						{ observable: observableTypes.subEntity, rel: 'third' }
-					]
-				},
-				subEntitySecondAnswer: {
-					observable: observableTypes.property, id: 'answer',
-					route: [{ observable: observableTypes.subEntity, rel: 'second' }]
-				},
-				primeLinkAnswer: {
-					observable: observableTypes.property, id: 'answer',
-					route: [{ observable: observableTypes.link, rel: 'prime' }]
-				},
-				primeLinkSubEntityAnswer: {
-					observable: observableTypes.property, id: 'answer',
-					route: [
-						{ observable: observableTypes.link, rel: 'prime' },
-						{ observable: observableTypes.subEntity, rel: 'forth' }
-					]
-				},
-			};
-			const observer = {};
-			state.addObservables(observer, observable);
-
-			fetch(state);
-			expect(observer.parentAnswer).to.be.undefined;
-			expect(observer.subEntityFirstAnswer).to.be.undefined;
-			expect(observer.subEntitySecondAnswer).to.be.undefined;
-			expect(observer.primeLinkAnswer).to.be.undefined;
-			expect(observer.subEntityFirstSubAnswer).to.be.undefined;
-			expect(observer.primeLinkSubEntityAnswer).to.be.undefined;
-			await state.allFetchesComplete;
-			expect(observer.parentAnswer).to.equal(42);
-			expect(observer.subEntityFirstAnswer).to.equal(43);
-			expect(observer.subEntityFirstSubAnswer).to.equal(44);
-			expect(observer.subEntitySecondAnswer).to.equal(45);
-			expect(observer.primeLinkAnswer).to.equal(46);
-			expect(observer.primeLinkSubEntityAnswer).to.equal(47);
 		});
 	});
 });
